Move routing root to App.js
